requirements.txt
pytest
webtest
bottle
cassandra-driver
celery
ipython
jupyter
